+++ react/sagas/index.js
@@ -1,11 +1,14 @@
 import { all } from 'redux-saga/effects';
-import { watchNewAssetRequest, watchShowNewAsset, watchNewChannelRequest, watchShowNewChannel, watchFileIsRequested } from './show';
+import { watchNewAssetRequest, watchNewChannelRequest } from './request';
+import { watchFileIsRequested } from './file';
+import { watchShowNewChannel } from './show_channel';
+import { watchShowNewAsset } from './show_asset';
 
 export default function* rootSaga () {
   yield all([
     watchNewAssetRequest(),
-    watchShowNewAsset(),
     watchNewChannelRequest(),
+    watchShowNewAsset(),
     watchShowNewChannel(),
     watchFileIsRequested(),
   ]);
